Add files via upload
The first version establishes the foundation of the hospital management web app. It connects to the ROOMSORTING.db database and allows users to add and view patients and beds. Patients are automatically assigned to available beds using simple logic. The site uses pure HTML and Python to maintain simplicity and clarity. This version’s goal is to demonstrate the basic data interaction and flow between Flask and SQLite.