Add files via upload
In this stage, the hospital web app becomes more user-friendly and professional. It adds a “Hospital Profile” page, options to search and delete patients, and overall improves the interface for clarity and usability. The user experience is enhanced with a cleaner structure and logical navigation between pages, better reflecting a real-world hospital management dashboard.